requirements.txt
botbuilder-core==4.16.1
botbuilder-integration-aiohttp==4.16.1
aiohttp>=3.9.0
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
